API

API

Logging

Stats

Logging

SolverCore.log_header — Function

log_header(colnames, coltypes)

Creates a header using the names in colnames formatted according to the types in coltypes. Uses internal formatting specification given by SolverCore.formats and default header translation given by SolverCore.default_headers.

Input:

colnames::Vector{Symbol}: Column names.

coltypes::Vector{DataType}: Column types.

Keyword arguments:

hdr_override::Dict{Symbol,String}: Overrides the default headers.

colsep::Int: Number of spaces between columns (Default: 2)

See also log_row.

SolverCore.log_row — Function

log_row(vals)

Creates a table row from the values on vals according to their types. Pass the names and types of vals to log_header for a logging table. Uses internal formatting specification given by SolverCore.formats.

To handle a missing value, add the type instead of the number:

@info log_row(Any[1.0, 1])
@info log_row(Any[Float64, Int])

Prints

[ Info:  1.0e+00       1
[ Info:        -       -

Keyword arguments:

colsep::Int: Number of spaces between columns (Default: 2)

Stats

SolverCore.GenericExecutionStats — Type

GenericExecutionStats(status, nlp; ...)

A GenericExecutionStats is a struct for storing output information of solvers. It contains the following fields:

status: Indicates the output of the solver. Use show_statuses() for the full list;

solution: The final approximation returned by the solver (default: []);

objective: The objective value at solution (default: Inf);

dual_feas: The dual feasibility norm at solution (default: Inf);

primal_feas: The primal feasibility norm at solution (default: 0.0 if uncontrained, Inf otherwise);

multipliers: The Lagrange multiplers wrt to the constraints (default: []);

multipliers_L: The Lagrange multiplers wrt to the lower bounds on the variables (default: []);

multipliers_U: The Lagrange multiplers wrt to the upper bounds on the variables (default: []);

iter: The number of iterations computed by the solver (default: -1);

elapsed_time: The elapsed time computed by the solver (default: Inf);

counters::NLPModels.NLSCounters: The Internal structure storing the number of functions evaluations;

solver_specific::Dict{Symbol,Any}: A solver specific dictionary.

The counters variable is a copy of nlp's counters, and status is mandatory on construction. All other variables can be input as keyword arguments.

Notice that GenericExecutionStats does not compute anything, it simply stores.

SolverCore.show_statuses — Function

show_statuses()

Show the list of available statuses to use with GenericExecutionStats.