Reference

Contents

Reference

Contents

Index

Index

SolverCore.AbstractSolver

SolverCore.AbstractSolverOutput

SolverCore.are_valid_parameters

SolverCore.can_solve_type

SolverCore.grid_search_tune

SolverCore.log_header

SolverCore.log_row

SolverCore.parameters

SolverCore.problem_types_handled

SolverCore.solve!

SolverCore.AbstractSolver — Type

AbstractSolver{T,S}

Base type for JSO-compliant solvers.

A solver must have three members:

initialized :: Bool, indicating whether the solver was initialized

params :: Dict, a dictionary of parameters for the solver

workspace, a named tuple with arrays used by the solver.

The type T is used for element types of the arrays, and S is used for the storage container type.

SolverCore.AbstractSolverOutput — Type

AbstractSolverOutput{T,S}

Base type for output of JSO-compliant solvers. An output must have at least the following:

status :: Symbol

solution

The type T is used for element types of the arrays, and S is used for the storage container type.

SolverCore.are_valid_parameters — Function

are_valid_parameters(solver, args...)

Return whether the parameters given in args are valid for solver. The order of the parameters must be the same as in parameters(solver).

Solvers should define

SolverCore.are_valid_parameters(::Type{Solver{T}}, arg1, arg2, ...) where T

SolverCore.can_solve_type — Method

can_solve_type(solver, type)

Check if the solver can solve problems of type. Call problem_types_handled for a list of problem types that the solver can solve.

SolverCore.grid_search_tune — Method

solver, results = grid_search_tune(SolverType, problems; kwargs...)

Simple tuning of solver SolverType by grid search, on problems, which should be iterable. The following keyword arguments are available:

success: A function to be applied on a solver output that returns whether the problem has terminated succesfully. Defaults to o -> o.status == :first_order.

costs: A vector of cost functions and penalties. Each element is a tuple of two elements. The first is a function to be applied to the output of the solver, and the second is the cost when the solver fails (see success above) or throws an error. Defaults to

[
  (o -> o.elapsed_time, 100.0),
  (o -> o.counters.neval_obj + o.counters.neval_cons, 1000),
  (o -> !success(o), 1),
]

which represent the total elapsed_time (with a penalty of 100.0 for failures); the number of objective and constraints functions evaluations (with a penalty of 1000 for failures); and the number of failures.

grid_length: The number of points in the ranges of the grid for continuous points.

solver_kwargs: Arguments to be passed to the solver. Note: use this to set the stopping parameters, but not the other parameters being optimize.

Any parameters accepted by the Solver: a range to be used instead of the default range.

The default ranges are based on the parameters types, and are as follows:

:real: linear range from :min to :max with grid_length points.

:log: logarithmic range from :min to :max with grid_length points. Computed by exp of linear range of log(:min) to log(:max).

:bool: either false or true.

:int: integer range from :min to :max.

SolverCore.log_header — Method

log_header(colnames, coltypes)

Creates a header using the names in colnames formatted according to the types in coltypes. Uses internal formatting specification given by SolverCore.formats and default header translation given by SolverCore.default_headers.

Input:

colnames::Vector{Symbol}: Column names.

coltypes::Vector{DataType}: Column types.

Keyword arguments:

hdr_override::Dict{Symbol,String}: Overrides the default headers.

colsep::Int: Number of spaces between columns (Default: 2)

See also log_row.

SolverCore.log_row — Method

log_row(vals)

Creates a table row from the values on vals according to their types. Pass the names and types of vals to log_header for a logging table. Uses internal formatting specification given by SolverCore.formats.

To handle a missing value, add the type instead of the number:

@info log_row(Any[1.0, 1])
@info log_row(Any[Float64, Int])

Prints

[ Info:  1.0e+00       1
[ Info:        -       -

Keyword arguments:

colsep::Int: Number of spaces between columns (Default: 2)

SolverCore.parameters — Function

named_tuple = parameters(solver)
named_tuple = parameters(SolverType)
named_tuple = parameters(SolverType{T})

Return the parameters of a solver, or of the type SolverType. You can specify the type T of the SolverType. The returned structure is a nested NamedTuple. Each key of named_tuple is the name of a parameter, and its value is a NamedTuple containing

default: The default value of the parameter.

type: The type of the parameter, such as Int, Float64, T, etc.

and possibly other values depending on the type. Some possibilies are:

scale: How to explore the domain

:linear: A continuous value within a range

:log: A positive continuous value that should be explored logarithmically (like 10⁻², 10⁻¹, 1, 10).

min: Minimum value.

max: Maximum value.

Solvers should define

SolverCore.parameters(::Type{Solver{T}}) where T

SolverCore.problem_types_handled — Method

problem_types_handled(solver)

List the problem types handled by the solver.

SolverCore.solve! — Function

output = solve!(solver, problem)

Solve problem with solver.